Added the Repository/Service/Controller for Brands

+++ src/SportsStore2.API/Models/User.cs
@@ -14,8 +14,8 @@ namespace SportsStore2.API.Models
         public string Name { get; set; }
         public string Surname { get; set; }
         public string Email { get; set; }
-        public int? HomeNo { get; set; }
-        public int? MobNo { get; set; }
+        public short? HomeNo { get; set; }
+        public short? MobNo { get; set; }
 
         public virtual ICollection<Address> Address { get; set; }
     }
